Move HashableObject to fixtures NS
